requirements.txt
fastapi==0.113.0
numpy==1.24.4
openai==1.43.1
pydantic==2.9.0
python-dotenv==1.0.1
